Fixes #3 - Add ability to export image
A new attribute named "FormatInfoAttribute" has been added to the project,
this attribute is used to add metadata about each image format in the
ImageFormat enum. This metadata includes things like a description of the
image format and the list of file extensions that are associated with the
image format. This metadata is then used at run-time for a couple of different
purposes.

The first is to generate the file filter string for the SaveFileDialog
instance that is used by the user to choose where to export the image to. The
file filter string is used to limit the "Save as type:" dropdown of the
SaveFileDialog to just the image formats that are contained in the
ImageFormat enum. A new class named "SaveFileDialogBuilder" has been created to
encapsulate the logic for building this SaveFileDialog instance, it uses the
new "ImageFormatEnumParser" class to parse out all the metadata in ImageFormat
in order to build the file filter string.

The second is to determine which format to export the image as when the user
closes the save dialog. The logic is to grab the extension from the file name
that the user selected, grab all the metadata from the ImageFormat, then find
the enum value of the image format that corresponds to the file extension. A
new ConvertAsync(string dotFile, string outputFile) method has been added to
DotFileImageConverter that encapsulates this logic.

A new "Save As..." menu item has been added to MainWindow.xaml that is used
by the user to initiate the image export process. This menu item can only
be clicked when a dot file is open, a new property named "DotFileOpened"
has been added to MainWindowViewModel to keep track of this. It merely
returns true or false based on whether or not the new "DotFilePath" property
is populated or not.

A click handler for the new menu item has been added to MainWindow that uses
the new SaveFileDialogBuilder to create a new SaveFileDialog instance and then
calls the new ConvertAsync() method added to DotFileImageConvert to export the
image.